+++ packages/dumb/src/index.ts
@@ -60,6 +60,8 @@ import CbgSliceSegment from './components/trends/cbg/cbg-slice/cbg-slice-segment
 import { NoDataLabel } from './components/trends/common/no-data-label/no-data-label'
 import { CbgSliceAnimated } from './components/trends/cbg/cbg-slice/cbg-slice-animated'
 import { BgPrefs } from './models/blood-glucose.model'
+import { XAxisLabels } from './components/trends/common/x-axis-labels/x-axis-labels'
+import { XAxisTicks } from './components/trends/common/x-axis-ticks/x-axis-ticks'
 
 export {
   AverageDailyDoseStat,
@@ -90,5 +92,7 @@ export {
   TerminalTable,
   Tooltip,
   TotalInsulinStat,
-  TotalCarbsStat
+  TotalCarbsStat,
+  XAxisLabels,
+  XAxisTicks
 }