+++ packages/yourloops/test/integration/pages/trends/trends-view-common.spec.tsx
@@ -31,6 +31,7 @@ import {
   checkDaysSelection,
   checkRangeSelection,
   checkSMBGTrendsStatsWidgetsTooltips,
+  checkTrendsLayout,
   checkTrendsStatsWidgetsTooltips,
   checkTrendsTidelineContainerTooltips,
   checkTrendsTimeInRangeStatsWidgets
@@ -45,7 +46,7 @@ import {
   checkTimeInRangeStatsTitle
 } from '../../assert/stats'
 import userEvent from '@testing-library/user-event'
-import { screen } from '@testing-library/react'
+import { act, screen } from '@testing-library/react'
 
 describe('Trends view for anyone', () => {
   beforeAll(() => {
@@ -68,7 +69,11 @@ describe('Trends view for anyone', () => {
       await checkRangeSelection()
       await checkDaysSelection()
 
-      await userEvent.click(screen.getByTestId('button-nav-back'))
+      checkTrendsLayout()
+
+      await act(async () => {
+        await userEvent.click(screen.getByTestId('button-nav-back'))
+      })
       expect(await screen.findByText('There is no CGM data for this time period :(')).toBeVisible()
     })
   })